FLUID-4209: Start of a component to render heading levels
This is the start of refactoring the tableOfContents component. The first step is to create a subcomponent to handle the rendering of each layer of the toc hierarchy. This is the first pass at that. Currently only the heading links get rendered. There still needs to be recursion down to subheadings.